--- server/controllers/__init__.py
@@ -1,6 +1,0 @@
-from server.controllers import animal_controller
-from server.controllers import auth_controller
-from server.controllers import cart_controller
-from server.controllers import home_controller
-from server.controllers import user_controller
-
