--- server/models/__init__.py
@@ -1,4 +1,0 @@
-from server.models import auth
-from server.models import user_info
-from server.models import animal
-from server.models import animal
