--- server/seeds/__init__.py
@@ -1,4 +1,0 @@
-from server.seeds import animal_s
-from server.seeds import auth_s
-from server.seeds import cart_s
-from server.seeds import user_info_s
